requirements.txt
configparser == 6.0.0
customtkinter == 5.2.1
gspread == 5.11.2
openpyxl == 3.1.2
packaging == 23.2
pandas == 2.1.1
pillow == 10.0.1
progress == 1.6
pytz == 2023.3.post1
requests == 2.31.0
xlwings == 0.30.12
